requirements.txt
torch
torchvision
torchaudio
openai-whisper
SpeechRecognition
PyQt6
numpy
pyaudio
